Adjust intersection filter types, add graduations
Unfortunately we have to add graduations to get the
labels below. I think it would be great to have labels so when people
screenshot the site the filter shows what is showing.

See docstring of intersection filter component for why I think
graduations should be multiples of 5. Happy to remove graduations
if anyone finds a way to still show the labels below.